pyproject.toml
[project]
name = "ethernaut-arena-green-agent"
version = "0.1.0"
description = "Ethernaut Arena Green Agent for AgentBeats platform"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # A2A dependencies
    "a2a-sdk[http-server]>=0.3.20",
    "pydantic>=2.12.5",
    "uvicorn>=0.38.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.1.1",

    # Ethernaut-specific dependencies
    "web3>=7.0.0",
    "loguru>=0.7.0",
    "aiohttp>=3.9.0",
    "py-solc-x>=2.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
